Replace genkeys.py with genkeys_new.py
